+++ xboot-fast/src/main/java/cn/exrick/xboot/modules/your/entity/Template.java
@@ -1,11 +1,10 @@
 package cn.exrick.xboot.modules.your.entity;
 
 import cn.exrick.xboot.base.XbootBaseEntity;
-import cn.exrick.xboot.modules.base.entity.Permission;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableName;
-import io.swagger.annotations.ApiModelProperty;
 import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import javax.persistence.Entity;
@@ -35,28 +34,28 @@ public class Template extends XbootBaseEntity {
     @ApiModelProperty(value = "内容")
     private String content;
     @ApiModelProperty(value = "分值")
-    private Integer score;
+    private Double score;
     @ApiModelProperty(value = "排序")
     private Integer sortOrder;
     @ApiModelProperty(value = "级别")
     private Integer level;
     @ApiModelProperty(value = "层级类型0标题，1内容")
-    private  Integer levelType;
+    private Integer levelType;
 
     @Transient
-    @TableField(exist=false)
+    @TableField(exist = false)
     @ApiModelProperty(value = "是否为上级")
     private Boolean isParent;
     @Transient
-    @TableField(exist=false)
+    @TableField(exist = false)
     @ApiModelProperty(value = "上级名称")
     private String parentTitle;
     @Transient
-    @TableField(exist=false)
+    @TableField(exist = false)
     @ApiModelProperty(value = "孩子")
     private List<Template> children;
     @Transient
-    @TableField(exist=false)
+    @TableField(exist = false)
     @ApiModelProperty(value = "分类名称")
     private String typeTitle;
